+++ src/player.h
@@ -279,7 +279,7 @@ void player_playback_stop (player_t *player);
 void player_playback_pause (player_t *player);
 void player_playback_seek (player_t *player, int value, player_pb_seek_t seek);
 void player_playback_seek_chapter (player_t *player, int value, int absolute);
-void player_playback_speed (player_t *player, int value);
+void player_playback_speed (player_t *player, float value);
 
 /***************************************************************************/
 /*                                                                         */
